# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Siuji TOEFL exam application.",
        "title": "Siuji API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/score-conversions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScoreConversion"
                ],
                "summary": "List score conversions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by listening, structure, or reading",
                        "name": "section_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.ScoreConversionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScoreConversion"
                ],
                "summary": "Create score conversion entry",
                "parameters": [
                    {
                        "description": "Conversion entry",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ScoreConversionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ScoreConversionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/score-conversions/bulk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Insert or update multiple entries at once, keyed by (section_type, correct_count).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScoreConversion"
                ],
                "summary": "Bulk upsert score conversions",
                "parameters": [
                    {
                        "description": "List of conversion entries",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BulkScoreConversionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.BulkScoreConversionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
        "/score-conversions/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScoreConversion"
                ],
                "summary": "Update score conversion entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Score conversion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Conversion entry",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ScoreConversionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ScoreConversionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ScoreConversion"
                ],
                "summary": "Delete score conversion entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Score conversion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseNoData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.BulkScoreConversionRequest": {
            "type": "object",
            "required": [
                "conversions"
            ],
            "properties": {
                "conversions": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.ScoreConversionRequest"
                    }
                }
            }
        },
        "model.BulkScoreConversionResponse": {
            "type": "object",
            "properties": {
                "total_saved": {
                    "type": "integer"
                }
            }
        },
...
        "model.ScoreConversionRequest": {
            "type": "object",
            "required": [
                "section_type"
            ],
            "properties": {
                "correct_count": {
                    "type": "integer",
                    "minimum": 0
                },
                "scaled_score": {
                    "type": "integer",
                    "minimum": 0
                },
                "section_type": {
                    "type": "string",
                    "enum": [
                        "listening",
                        "structure",
                        "reading"
                    ]
                }
            }
        },
        "model.ScoreConversionResponse": {
            "type": "object",
            "properties": {
                "correct_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "scaled_score": {
                    "type": "integer"
                },
                "section_type": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.ResponseNoData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response_code": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "TempToken": {
            "description": "Type \"Bearer\" followed by a space and temp JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}